Separated EditTaskPanel into two components, refactoring
