+++ app/static/app/js/css/EditTaskForm.scss
@@ -1,0 +1,3 @@
+.edit-task-form{
+
+}
